requirements.txt
st-gsheets-connection
streamlit
pandas
numpy
plotly.express
matplotlib.pyplot
seaborn